blink-cam/requirements.txt
# Blink Camera Face Recognition System
opencv-python>=4.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
requests>=2.31.0
#face_recognition>=1.3.0
ultralytics>=8.0.0
